+++ c1/Libft/ft_memcmp.c
@@ -6,11 +6,13 @@
 /*   By: sersanch <sersanch@student.42barcelona.    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/09/15 13:34:53 by sersanch          #+#    #+#             */
-/*   Updated: 2022/09/16 15:02:14 by sersanch         ###   ########.fr       */
+/*   Updated: 2022/09/16 15:29:29 by sersanch         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-int	memcmp(const void *s1, const void *s2, unsigned int n)
+#include "libft.h"
+
+int	memcmp(const void *s1, const void *s2, size_t n)
 {
 	int		i;
 	char	*s1c;
