+++ c1/Libft/ft_strnstr.c
@@ -6,13 +6,15 @@
 /*   By: sersanch <sersanch@student.42barcelona.    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/09/15 12:44:09 by sersanch          #+#    #+#             */
-/*   Updated: 2022/09/16 14:55:41 by sersanch         ###   ########.fr       */
+/*   Updated: 2022/09/16 15:32:16 by sersanch         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-char	*ft_strnstr(const char *str, const char *to_find, unsigned int len)
+#include "libft.h"
+
+char	*ft_strnstr(const char *str, const char *to_find, size_t len)
 {
-	unsigned int	i;
+	size_t			i;
 	int				j;
 	char			*pnt_first;
